fix: keep JavaScript invocation service within package limit
